+++ src/test/java/nextstep/subway/station/StationAcceptanceTest.java
@@ -4,9 +4,12 @@ import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import io.restassured.response.ValidatableResponse;
+import nextstep.subway.BaseAcceptanceTest;
+import nextstep.subway.domain.StationRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
@@ -20,16 +23,11 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
 @DisplayName("지하철역 관련 기능")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class StationAcceptanceTest {
-    @LocalServerPort
-    int port;
+public class StationAcceptanceTest extends BaseAcceptanceTest {
 
     @BeforeEach
-    public void setUp() {
-        if (RestAssured.port == RestAssured.UNDEFINED_PORT) {
-            RestAssured.port = port;
-        }
+    void setup(@Autowired StationRepository stationRepository) {
+        stationRepository.deleteAll();
     }
 
     /**
@@ -134,8 +132,8 @@ public class StationAcceptanceTest {
                           .then().log().all();
     }
 
-    private static void assertStatusCode(ExtractableResponse<Response> response, HttpStatus badRequest) {
-        assertThat(response.statusCode()).isEqualTo(badRequest.value());
+    private static void assertStatusCode(ExtractableResponse<Response> response, HttpStatus httpStatus) {
+        assertThat(response.statusCode()).isEqualTo(httpStatus.value());
     }
 
     private static List<String> toStationNames(ExtractableResponse<Response> response) {
